+++ pypgx/api/core.py
@@ -481,7 +481,7 @@ def get_priority(gene, phenotype):
 
     return df[i].Priority.values[0]
 
-def get_ref_allele(gene, assembly='GRCh37'):
+def get_ref_allele(gene):
     """
     Get the reference allele for target gene.
 
@@ -1372,7 +1372,7 @@ def sort_alleles(
         impacts = [get_variant_impact(x) for x in core_variants]
         impacts = [x for x in impacts if x]
         c = len(impacts) * -1
-        d = allele == get_ref_allele(gene, assembly=assembly)
+        d = allele == get_ref_allele(gene)
         return (a, b, c, d)
 
     def func2(allele):